pyproject.toml
[tool.poetry]
name = "eth-transaction-exporter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
web3 = "^6.0.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pandas = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"